Fill width with spaces and return to line start
The previous approach of counting line characters and filling spaces
at the end to meet the requested width is complicated because not
all characters occupy one character cell when displayed. For
instance:

 - Tab characters can be various widths based on tabs(1) settings.
 - Zero-width and non-visible characters like escape codes.
 - Wide characters that take up multiple character cells.

It may be possible to work through all this to determine visible
character width but it'd take a while.

The downside to the new approach is that the output size is greatly
increased and the \r and space characters may cause problems when
piping output to filter commands or copy/pasting.